refactor: break up openai
