+++ rig-core/examples/agent_autonomous.rs
@@ -1,4 +1,4 @@
-use rig::providers::openai::Client;
+use rig::providers::openai::client::Client;
 use schemars::JsonSchema;
 use std::env;
