+++ rig-core/examples/hyperbolic_image_generation.rs
@@ -1,5 +1,5 @@
 use rig::image_generation::ImageGenerationModel;
-use rig::providers::{hyperbolic, openai};
+use rig::providers::hyperbolic;
 use std::env::args;
 use std::fs::File;
 use std::io::Write;
